+++ R3nzSkin/offsets.hpp
@@ -22,31 +22,31 @@ namespace offsets {
 	};
 
 	namespace global {
-		inline std::uint32_t Player{ 0 };
-		inline std::uint32_t ChampionManager{ 0 };
-		inline std::uint32_t Riot__g_window{ 0 };
-		inline std::uint32_t ManagerTemplate_AIMinionClient_{ 0 };
-		inline std::uint32_t ManagerTemplate_AIHero_{ 0 };
-		inline std::uint32_t ManagerTemplate_AITurret_{ 0 };
-		inline std::uint32_t GameClient{ 0 };
-		inline std::uint32_t retSpoofGadget{ 0 };
+		inline std::uint64_t Player{ 0 };
+		inline std::uint64_t ChampionManager{ 0 };
+		inline std::uint64_t Riot__g_window{ 0 };
+		inline std::uint64_t ManagerTemplate_AIMinionClient_{ 0 };
+		inline std::uint64_t ManagerTemplate_AIHero_{ 0 };
+		inline std::uint64_t ManagerTemplate_AITurret_{ 0 };
+		inline std::uint64_t GameClient{ 0 };
+		inline std::uint64_t retSpoofGadget{ 0 };
 	};
 
 	namespace AIBaseCommon {
-		inline std::uint32_t CharacterDataStack{ 0 };
-		inline std::uint32_t SkinId{ 0 };
+		inline std::uint64_t CharacterDataStack{ 0 };
+		inline std::uint64_t SkinId{ 0 };
 	};
 
 	namespace MaterialRegistry {
-		inline std::uint32_t D3DDevice{ 0 };
-		inline std::uint32_t SwapChain{ 0 };
+		inline std::uint64_t D3DDevice{ 0 };
+		inline std::uint64_t SwapChain{ 0 };
 	};
 
 	namespace functions {
-		inline std::uint32_t Riot__Renderer__MaterialRegistry__GetSingletonPtr{ 0 };
-		inline std::uint32_t translateString_UNSAFE_DONOTUSE{ 0 };
-		inline std::uint32_t CharacterDataStack__Push{ 0 };
-		inline std::uint32_t CharacterDataStack__Update{ 0 };
-		inline std::uint32_t GetGoldRedirectTarget{ 0 };
+		inline std::uint64_t Riot__Renderer__MaterialRegistry__GetSingletonPtr{ 0 };
+		inline std::uint64_t translateString_UNSAFE_DONOTUSE{ 0 };
+		inline std::uint64_t CharacterDataStack__Push{ 0 };
+		inline std::uint64_t CharacterDataStack__Update{ 0 };
+		inline std::uint64_t GetGoldRedirectTarget{ 0 };
 	};
 };
